fixed  trans Opt scoring
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk/moses@92 1f5c12ca-751b-0410-a591-d2e778427230

+++ src/StaticData.h
@@ -111,7 +111,7 @@ public:
 	{
 		return m_weightWordPenalty;
 	}
-	unsigned int GetMaxHypoStackSize() const
+	size_t GetMaxHypoStackSize() const
 	{
 		return m_maxHypoStackSize;
 	}
